paginated query for data class children

+++ ModelCatalogueCorePluginTestApp/grails-app/controllers/org/modelcatalogue/core/ChildrenDataCommand.groovy
@@ -3,8 +3,8 @@ import grails.validation.Validateable
 
 @Validateable
 class ChildrenDataCommand {
-    Long offset
-    Long max
+    int offset
+    int max
     static constraints = {
         offset nullable: false
         max nullable: false
